Rework command return value
The type for the return value of a Command is now parameterized, instead
of being only a boolean. Most commands are updated to still return
booleans, but a couple return strings now.

In order to tell if a command executed but failed, CommandException is
introduced. Commands now throw this when they fail, instead of just
returning a false value. Shell looks for this exception and reports its
message to the actor, keeping the prior apperance of command failure.

Therefore, a CommandResult knows command execution is successful when
there is no CommandException or CommandFactoryException recorded for
execution or build failure, respectively.