Add capabilities to scripts
Scripts run as commands have more to work with.

- The extension housing the script is bound as "extensionId".
- The new runAsOwner() function lets a script switch to running as its
owner, instead of the calling player. This allows scripts to do things
that the original caller cannot. BE WARNED.
- The new getAttr() function lets a script get a full-blown Attr object
from a thing in the caller's universe. This does check whether the
caller has READ permission on the thing. (This might become a command in
the future, but it is nice to provide an object instead of a string.)
- The new failure() function makes it easy to build a CommandException
for when a script needs to fail.

Finally, the executeCommand function is shortened to just exec.