Start command testing
A few unit tests are now in place for some commands, particularly those
with permission checks. The new AbstractCommandTest class helps with
writing this tests.

Code changes driven by the tests:

- InspectCommand, LookCommand, MoveCommand, and TeleportCommand now
return a string.
- For testability, TeleportCommand allows for setting the follow-on
LookCommand it uses.
- TakeCommand now sets the location of the taken thing to null. (This
was probably a bug.)

The new tests use Mockito, so that is now included as a test dependency.
Calling getClass() on a mock gives you the "wrong" answer, so the new
TestabilityUtils.nonmock class helps with that. TakeCommand uses nonmock
so that it can be tested with mock things.

PMD checks are now extended to unit tests, so some other existing tests
are fixed up as well.