Update message handling
Messages to actors are no longer just strings, but objects that record
the sender and the message content. A message also has a style that
controls how it is formatted. There are several pre-defined styles,
including some for certain commands like pose, say, and whisper (so
those commands no longer need to define them).

From this:

- HelpCommand now returns a string.
- The message written out by Shell for a CommandException is now
prefixed with an exclamation point.

GroundsScript.runAsOwner is fixed - it was doing the opposite of what it
should do. To make things clearer, GroundsScript now distinguishes the
original player calling the script ("caller") from the player running
the script ("runner", which may end up being the owner). A script can
now get the original caller's name and also send a message to any
player, not just the caller.