Starting attempt to sandbox scripts
Scripts should not be trusted code, and the groovy-sandbox project aims
to have scripts run in a sandbox so that they can't, say, call
System.exit or use reflection or do other things. I tried to integrate
it, but it's not working out.

- The sandbox library requires Groovy 2.
- For some reason, in my tests, my (chat) script needs java.lang.Class,
which is a bad idea to allow.
- The script also needs not just String, but arrays of String, which is
a separate class.

At the moment, sandboxing seems intractable.