Add ServerCommand and WhoCommand, add actor controls
The new WhoCommand gives a list of the actors and their current players
in the current universe. It will be refined further, for sure.

WhoCommand extends the new ServerCommand type, which is for commands
that can work with the Server object running a multi-user game.

Server now keeps track of some state:

* A set of connected actors. This allows the server to block multiple
connections by a single actor. This also might change later, so that an
actor can connect as multiple players (characters).
* The shells that are open for actors. This enables WhoCommand to report
on the player that each actor is currently playing as. This mapping is
weak, so that if an actor disconnects uncleanly, their shell doesn't
linger.

Because ServerCommand needs a server reference, CommandFactory needs to
receive that reference so it can make ServerCommand instances.
CommandFactory knows how to build ServerCommands now, passing the server
to their "newCommand" method.

This caused trouble for the singleton CommandExecutor class - sometimes
there is a Server, sometimes not (when in single-user mode). So, that
static singleton is replaced with one that is created by either Server
or SingleUser. The singleton can still only be set once, so the
guarantees of having just one executor remain.