Replace Shell.preprocess with transforms
The crappy preprocess method in Shell is gone, replaced by functions
that transform tokenized command lines into different ones. The
transforms are defined in CommandExecutor next to the command name
mappings, where they all belong (at the moment).

The current transforms are the same as the replaced preprocessing:
support for "OOC" as a command name, and for ':' as a shortcut for
posing.