Add chat system
Finally! The command script and YAML files in etc/ext/chat establish a
chat system based on scripted commands. The chat system is oriented
around named channels. Players may join and leave channels at will, and
say things in channels they are joined to. Administrators can create and
delete channels.

There are still things to fix in the system, and to add, but it is
functioning at a minimal level. Most recent work in this project has
been towards making chat possible, and now is a good starting point for
a complete system.