Add DestroyCommand
The new DestroyCommand lets you destroy things that have been built.
Conditions apply to some destructions, such as not being able to destroy
a place that is still linked or a player still being played by some
actor. The command doesn't yet know how to destroy a universe, which is
probably for the best.

Any contents of a destroyed thing are sent off to a new "lost and found"
place for the thing's universe. From now on, when a universe is built,
it gets this place as well as an origin.

A pretty complete unit test is in place for DestroyCommand, which
spurred on some improvements to AbstractCommandTest. A BuildCommand unit
test is also kicked off, but coverage is low right now. (BuildCommand
now throws a PermissionException when called by a non-wizard, not just
CommandException.)

closes #6