Track IP addresses for actors
Actor objects and the actor database now store the most recent IP
address observed for each actor.

The server finds the IP address from the channel session (as before, in
order to block root logins from anywhere but localhost) and stores it
for the actor before spawning their shell. Shell also stores this IP
address, and lets it appear in the welcome banner.

In single-user mode, 127.0.0.1 is assumed and passed to Shell.

WhoCommand is upgraded in a couple of ways. First, it supports being
called with a single argument "all" to list all actors, not just those
with open shells. Second, the listing includes the current / most recent
IP address for each actor, if known. The sorting of connected actors
should also be fixed.

closes #16