Add actor preferences, including ansi
An actor now has a set of preferences (name / value pairs) available.
The first use of this is to define an "ansi" preference that indicates
whether the actor wishes for game output to include ANSI escape codes
for coloration. The Shell class disables ANSI based on the preference
value.

To support this, both Server and SingleUser now take care to load actor
preference information from the actor database upon login. And because
of that, SingleUser now creates a new actor database containing only the
root actor if it doesn't exist.