Add home concept, and handle unresolvable places
The new HOME command lets players set a home place for themselves and
also teleport to it.

As part of this change, I also tackled the issue of what happens when a
player (or thing) has a location or home set but it can't be found. In
most cases, this will cause a command to fail, but some commands can
work around it.

This work starts to address #7. When a player has a home, on
disconnection they can be teleported there.