Implement autosave
The game server now runs a scheduled thread that periodically saves the
current universe to its file. The current universe's file is remembered
by the Universe class and maintained by the various commands that work
with it and the current universe, along with the Main class that sets
things up in the first place.

The key class for autosave is AutosaveRunnable, which is what the
scheduled thread actually calls upon. It works by executing a custom
AutosaveCommand which performs the save operation. The reason for a
command is so that the work is queued along with all other commands that
potentially change game state. Since there is only one thread processing
commands, this ensures that the universe's state doesn't change during
the save process (and synchronization isn't an issue).

closes #8