Implement actor locking
New actor lock and unlock commands enable actors to be locked until some
time, or unlocked. While an actor is locked, they may not access the
server, although if they are already connected, they are not booted.

Only GOD can lock and unlock actors, but it looks like these commands
and some other actor commands should be made available to thaumaturges
as well.

The locked actor doesn't see the message that they are locked out, which
would be better than a mystery failure. Slightly.