Add BootActorCommand
The new ACTOR BOOT command makes it possible to terminate the shells of
an actor, disconnecting them from the server. The command finds all
shells for the actor, even though right now an actor can only be
connected once; that will change in the future.

The Shell class gets a new terminate method to make this possible for
the command. The implementation of this requires the shell to have a
reference to its own Future, which might be a little out of the
ordinary, but it seems OK for now.

Since BootActorCommand must be a ServerCommand, the weird ActorCommand
composite has become one as well, and it can no longer use a static
command factory internally for building actor-related commands. Another
reason to maybe rework how the actor commands are bundled together
(along with lack of support for help text).

The code for clearing expired actor locks is also improved so that it
doesn't happen all the time, but only when there is an unlock time to
clear.

closes #10