Switch MessageEmitter to work with players
MessageEmitter, which is responsible for emitting messages to users
through their shells, now works by player instead of by actor. This work
is building towards allowing a single actor to log in as multiple
players, so that messages are routed to the correct shells for that
actor depending on the relevant player. Upcoming changes will move the
message queues from Actor into Player, which will fully enable this.

Since a MessageEmitter is now scoped to a player, they are no longer
created once when an actor is logging in, but instead after they a)
select an initial player and b) again if/when they switch to a different
player. Shell is responsible for starting and stopping emitters. It
currently uses the same executor that runs shells themselves.

A subtle bug is also fixed here. When loading a new universe,
particularly in single-user mode, the current actor for GOD was not
preserved when the new universe included GOD as a thing. This prevented
messages from being emitted to GOD with the MessageEmitter changes,
since GOD would lack a current actor. LoadCommand now transfers the
actor to the GOD Player object in the loaded universe.

This work is towards implementation of #15