Support seizure, and things held by things
The TAKE command now explicitly allows ADEPTS and
THAUMATURGES to take things from other players ("seize").

This drives the need to know who is holding something, if
anyone, and so the location attribute for a thing may now
be a player to indicate direct possession. (Before, it was
just null / empty). The changes here go a little further,
since now a Place or Player may be a valid location, and
allow a Thing to be the location of another Thing. This can
represent nested objects, e.g., a player holding a rucksack
holding a gem. The TAKE command is smart enough to figure
this out.

A remaining limitation on locations enforced by the game is
that a player's location must be a Place. This does prevent
locating a player "in" a thing like a vehicle, for now.