Remove current player from actor
Actor no longer keeps track of the current player. It no
longer makes sense since an actor may play as multiple
players at once.

This led to an issue with SwitchPlayerCommand, which no
longer had a way to convey back to the Shell who the new
player is. To address this, SwitchPlayerCommand now has a
getNewPlayer method to provide the player the command
switched the actor to.

To let Shell see this, CommandResult was modified so that
it no longer stores just the type of the executed command,
but the command itself. Shell retrieves the player from the
executed command.