Use universe origin for player stowage
When a player leaves the server or switches to a new player,
they are relocated to the current universe's origin
("stowed"), which effectively removes their presence from
the running game. When the player is used again, they are
"brought out" to their home, or left in the current
universe's origin if they lack a home.

closes #7