Fix player auto-selection
With this changeset, if the single player permitted for an
actor is occupied, that player is not auto-selected when the
actor logs in. The actor sees the usual player selection
prompt, where they could wait until their player is no
longer occupied, or exit the game. (It might be nice in the
future to indicate in the listing which player(s) are
already occupied.)

The actor field of the Player class is now manipulated in a
thread-safe manner. This avoids race conditions where, for
example, two actors could still manage to occupy the same
player.

closes #19