Implement safe saves
The current universe is now saved in a "safe" manner by
first saving it to a temporary file, and then copying that
file's contents to the correct file. This way, if there is
a failure in the middle of saving, the original file is not
corrupted.

Safe saving happens for autosave, and when running the SAVE
command without a file argument (which is now legal).