Support help text for scripted commands
A Script, which contains code to be executed for a command,
now also may include a resource bundle of help text. The
HELP command checks if it is being asked about a scripted
command (by looking for that initial dollar sign) and, if
so, uses its resource bundle for help text.

Getting this to work well involved a good refactoring of
how scripted commands are constructed.

* ScriptFactory now looks for a script owner instead of
  ScriptedCommand, so the factory no longer needs to be
  passed an owner argument.
* The ScriptedCommand constructor and newCommand method
  no longer have arguments for the script's extension,
  because the Script object alread has it.
* The logic in CommandFactory for locating an attribute for
  a scripted command is exposed for other classes to use,
  such as HelpCommand.

Help text is added to the $chat command. Because attribute
values must always be text, it's a little tricky to express
the help text here; you can't just make a nested YAML
structure. However, because of the recent changes to Attr,
values here can be YAML strings at least, instead of long
JSON strings without newlines. That's enough to make
authoring help text OK.