Preserve GOD across universe loads
The handling of the GOD player when loading a new universe
is reworked. Usually GOD will be saved as a player in a
universe, but that player, when loaded, is a different
object than the current universe's GOD player. Before this
change, on load, GOD's actor was transferred from the old
GOD to the new one so that messages would still be
delivered, among other things. But this doesn't cover all
cases, such as SAY and POSE - there, messages were getting
sent correctly to the new GOD, but the message emitter was
hooked to the old GOD player. It's kind of confusing.

Anyway, I changed what happens to GOD during load. Now, if
the loaded universe has a GOD player, it is swapped out in
the loaded universe for the current one or the Player.GOD
singleton, which at this point should be one and the same.
The remaining player is then relocated to the former
location of the removed GOD player.