Use LoadCommand for initial universe load
When the game is started with the --universeFile option, the
file is loaded now by the Server class and not by the Main
class. This allows LoadCommand to always be used, ensuring
that the state-related work the command does happens
consistently.

This does remove the ability to load a universe initially
for single-user mode.