Implement listener attributes
Extensions may now have special listener attributes that
hold scripts to run in response to events. A listener
attribute has a name starting with a caret, and must be an
ATTRLIST type with at least a "scriptContent" attribute
holding the script to run in response. (The script is in
Groovy, as usual).

CommandExecutor now holds an event bus. Commands may record
events for interesting occurrences when they execute, and
then CommandCallable posts the events to the event bus. Each
command can name a player who the event is about, as well as
a payload object with additional information. The payload
must be able to be converted to JSON, so basic Java maps and
lists work fine.

Extensions themselves subscribe to the command event bus.
They are automatically subscribed when built, and
unsubscribed when destroyed. Also, when a universe is
loaded, LoadCommand unregisters all extensions from the
previous universe and registers all extensions from the
loaded one.

When an extension receives an event, it calls each of its
listener attributes' scripts, passing the event payload JSON
string as the sole argument. If an attribute has an
"eventType" attribute, then the event class name must match
that attribute's value for the script to be called. This
avoids compiling the script for listeners who don't care
about those events.

Each script execution is formed into a ScriptedCommand and
submitted to the command executor. So, the executions do not
occur in the same thread as the command which originally
posted the triggering event. (This means that the command
event bus can stay synchronous, for now.)

The actor for each of these scripted commands is a new
"internal" actor, since no actual actor is responsible for
the command execution. The player for each of the commands
is the extension itself.

Accordingly, Extension has been made a subclass of Player,
so that it can have roles and be responsible for commands.
This is a little odd, but it does replace the awkward need
to define a hidden "bot" player for an extension to serve as
its owner / holder of appropriate roles for it to run
scripts under. Extensions and scripts, therefore, no longer
require owners. A script can opt to run as its extension now
instead of its owner.

GroundsScript has a new parseEventPayload convenience method
for listener scripts to easily parse JSON payloads. The
former runAsOwner method has been swapped for runAsExtension
since the owner concept for scripts has been removed.

The first defined command event is TeleportArrivalEvent,
which is posted by TeleportCommand and MoveCommand to signal
that a player has arrived somewhere. Certainly more events
are to come.