Add localized listener attributes
A listener attribute on an extension may now declare itself
as localized. This means that it is only triggered for
events which occur in the attribute's extension's location.
Localized attributes are good for performance.

An example extension with listener attribute is provided;
whenever a player enters the extension's location, it poses
that a magic fiddle plays for the arriving player.