Auto-join guest players to #guest chat
The chat system extension gains a listener attribute that
adds any new guest player to the #guest channel. The
listener does this when a guest player is yoinked anywhere;
this happens when that player is first created and brought
out to the guest home of the universe. It could be run again
if the same guest is yoinked again, but that's harmless.

This highlights the need to automatically remove a guest
from #guest when they log off and their player is destroyed.
However, destruction events don't exist yet, and that should
happen for any player being destroyed, so it's a bigger bit
of work to do.

closes #24