Add visibility and join controls for chat channels
A chat channel now supports expressions that control whether
a player may see (visibility) or join (joinability) the
channel. When the expressions are not set, then anyone may
see or join the channel.

Each expression is written in Groovy and evaluated against
the channel attribute and the name of the script caller. So
far, I've only tested the trivial cases of "true" and
"false", so more work may be needed.

The Attr and GroundsScript classes gained some new useful
methods to help with general improvements in the chat
scripts.

This goes a long way towards resolving #22