Start supporting security manager use
A Java security policy is now available to help run Grounds
under a security manager. The policy permits everything for
the core application but severely restricts softcode (Groovy
scripts). The permissions granted to softcode were found
empirically, so it's pretty likely more will need to be
added later.

The policy requires setting the property jar.dir to the path
where the Grounds JAR is located.

The run.sh helps to run Grounds with or without a security
manager. If run with `--secure` as the first option, it sets
the jar.dir system property and the other properties needed
to enable a security manager with the policy. The script
looks for the JAR file and the policy file somewhere in its
own directory or below.

Somehow it turns out that even with the policy in place,
scripts may still successfully call System.exit. To counter
this, GroundsSecurityManager blocks VM exits completely.
This does block them for the core application as well, but
fortunately it doesn't need it anyway.

Finally, the Jackson ObjectMapper used to convert attributes
into JSON is modified so that it does not try to circumvent
Java access modifiers. This is necessary for it to work
under a security manager. Again, it's likely that other
changes like this will be discovered.

These changes are the foundation for resolving #3.