Define Grounds-specific permissions
New permissions guard critical core features from
unauthorized code, meaning scripts.

* ActorDatabasePermission guards all operations on the actor
  database.
* CommandExecutorPermission guards most operations on
  command executors, especially the singleton used by the
  game
* UniversePermission guards operations on the current
  universe and on saving universes.

At this time, scripts should not need any of the above
permissions to be granted in order to function. Instead,
they need to use GroundsScript methods as their gateways.
In particular, the sendMessageTo and exec methods of
GroundsScript run as privileged actions now, in order to
permit script threads to escalate their own permissions as
necessary.

At this point, the softcode security model is in place, but
there are bound to be updates and improvements later.

closes #3