Fix ANSI background handling, augment LookCommand
AnsiUtils now properly defaults background colors by only setting back
to the default background if a background color was set in the first
place. This avoids erroneous output.

Also, the LookCommand now emits the ID of the current location when the
actor's showIds preference is true. This lets the player know what to
teleport to later if desired.