Improve line breaks, especially for ANSI
LineBreaker is now ANSI-aware, so that lines with ANSI escape codes are
not broken sooner than necessary. It also does a better job of trimming
off trailing spaces on each line, not just when needed to make the line
short enough for the terminal width.

MessageEmiter is tweaked so that the updated line breaking works
correctly: each message is styled first before line breaks are added.

AnsiString receives some enhancements to help:

* toStrippedString returns the string contents stripped of escape codes
* equals and hashCode enable comparisons for testing
* isEmpty and substring are added for convenience