Allow some commands to resolve non-present players
The player argument for the following commands may now refer to a player
or location that is not in the same location as the player issuing the
command.

* MuteCommand (player)
* UnmuteCommand (player)
* TeleportCommand (location)

SendMailCommand and PageCommand already work the same way, and they are
refactored to use the common, updated argument resolver code used for
the other commands.

Bonus: You can no longer mute yourself.