Begin telnet support
Grounds may now run under a telnet server, adapted from the one included
in JLine 3, as an alternative to SSH.

* Telnetd is an adaptation of the Telnet builtin class from JLine. It's
much easier to use, and provides more information (all connection data)
to shells.
* SSH-specific parts of PasswordAuthenticator are refactored out to
SshdPasswordAuthenticator. The now neutral PasswordAuthenticator is
reused for authenticating over telnet as well.
* SSH-specific parts of Server are refactored out to SshServer. The now
neutral Server is reused as the basis for the new TelnetServer.
* The new `--telnet` option enables telnet. Grounds cannot (yet?) listen
on both SSH and telnet.

Important bits and pieces:

* The new hidePrompt preference lets an actor suppress prompts for
input; this reduces noise in Atlantis, which has a dedicated input pane.
* The shell Runnable run by Server is now a Callable that returns the
shell exit code. (Nothing uses that at the moment.)