Fix NPEs in WhoCommand and BootActorCommand
If an actor hasn't selected a player yet, their shell's player was
returned as null. This was causing WhoCommand to throw an NPE, since it
expected every open shell to have a player. The shell player is now
returned as an optional, and WhoCommand checks if a player is present.

BootActorCommand is updated to avoid a similar NPE. It also should once
again terminate all actor shells when no specific player for an actor is
targeted.