Notify when a player enters or exits a location
The TELEPORT (and therefore also MOVE) command emits enter and exit
messages to other players when a player enters or exits their location,
respectively.

The code for emitting these messages is refactored out of Shell and into
Command. Accordingly, the YOINK command takes on the responsibility of
notifying players when another player arrives or departs (via
login/logout or yoinking).