Build guest players using BuildCommand
Guest players are built by Shell using a build command instead of
directly. This closes a hole where changes to the universe could happen
outside of a command.

Accompanying changes:

* BuildCommand returns the string ID of the built thing instead of true.
* GOD can execute BuildCommand no matter where that player is, even if
it's nowhere. If GOD isn't in a findable location, new things are placed
at the current universe's origin.
* The CommandExecutor used by a Shell is now passed in on construction,
instead of always being the singleton instance. This will help in
writing upcoming tests. Server and Shell pass in the singleton.

Also an important change: type safety is improved for command execution.
This is primarily aimed at execution of commands that are built in code,
as opposed to those parsed from player input.

* CommandCallable now has a type parameter to match the return type of
the CommandResult produced by the command it calls.
* CommandExecutor enforces that the return type of a command submitted
to it matches the return type of the CommandResult.