Simplify CommandCallable
CommandCallable no longer can build commands itself, they must be
supplied to it. This radically simplifies the class.

Users of CommandCallable are updated so that they themselves construct
commands instead. This is a bit of an un-refactoring, but seems like a
better splitting of concerns. In particular, CommandExecutor.submit() is
now directly responsible for command construction.