Start API support
An ApiMethod holds logic for an API call by some method name. It takes
in a JSON RPC request and returns a JSON RPC response.

ApiMethodFactory provides ApiMethod instances. An ApiMethod instance is
stateless, because state is passed to it when called via an
ApiMethodContext object.

An ApiMethodContext holds the actor, running player, and other
information needed for an API method to understand the context of its
execution - for example, who ultimately caused the method to be called.

The idea is that a Grounds "plugin" will perform an API call. A plugin
doesn't call out of nowhere - for now - but instead because the plugin
is running a command some player is executing. So, an API call must pass
a "plugin call ID" to provide context around that command execution. The
call ID is associated with that context in the PluginCallTracker.

An API call must pass a plugin call ID using a request parameter. The
name of that parameter is in ApiRequestParameters.

An ApiHandler handles a single API call received through a ByteChannel.
It is response for parsing a JsonRpcRequest, validating that it has a
JSON RPC ID and plugin call ID, creating an ApiMethodContext from the
tracked plugin call information, finding and calling an ApiMethod, and
sending results back as a JsonRpcResponse to the channel.

ExecCommand is an initial ApiMethod implementation. It's not tested.