Add ApiServer
ApiServer listens on a domain socket for API calls, and passes them to
ApiHandler objects. Both single-user and multi-user mode now start an
API server if a domain socket path is configured for it.

Using Control-C to quit the server doesn't cause the ApiServer shutdown
process to occur at this time, which means the domain socket file won't
be deleted. ApiServer can't start again in a subsequent execution until
that file is deleted.