Add PluginCall support
Existing code for running scripted commands is copied and reworked to
support calling plugins.

PluginCall represents a call that Grounds makes to a plugin. This
includes a path to a plugin executable, a JSON-RPC method, a help text
resource bundle (like for scripts), the extension which is calling the
plugin on behalf of a player, and the plugin call tracker for the API
server.

PluginCallFactory builds PluginCall instances from extension "plugin
call" attributes. The process is a lot like building Scripts, by looking
for specific attributes in the list value of plugin call attributes. If
building a PluginCall fails, the factory throws a
PluginCallFactoryException.

PluginCallCommand is a command run by a player which calls a plugin.
Again, this works much like ScriptedCommand. Most of the work is done by
PluginCallable.

PluginCallable makes a call to a plugin and handles its response. For
now, it only supports spawning a new process and executing a plugin, but
in the future, it will be able to instead send a request to a running
plugin server process. Regardless, the plugin receives a JSON-RPC
request on standard input, and is expected to return a JSON-RPC response
on standard output. The result string from the plugin call is the result
of the callable.

PluginCallable also sets up tracking of its plugin call in the API
server and untracks when the call is complete.