Add asExtension parameter to ExecCommand
This causes the executed command to be run as the extension for the
plugin call that is calling the API method, instead of the original
caller. The extension may have permissions beyond what the caller has,
so a plugin can use that to execute more commands.

This replaces the "runAsExtension" concept from scripting. Instead of
making a blanket change to who is "running" the plugin call, instead a
plugin will specify for each API method whether to run as the extension.
(The only other upcoming API methods that need this are for sending
messages.)