Implement chat channel visibility / joinability by role
Chat channels may now list roles that are required to see and to join
them. The new $chatadmin commands set_visibility and set_joinability
enable chat admins to set roles for channels.

This also adds a $chatadmin inspect command to list all of the details
of a channel, including the roles lists.