Add ArrivalEvent and DepartureEvent
These new events are posted when a player (including a guest) logs in or
logs out. The events just point to the player, and no location or any
other payload.

The events are also posted when an actor switches players; the
switched-from departs, and the switched-to arrives.

Unlike all previous events, these two aren't tied to a command. Instead,
Shell posts them.