Support markup in record and tabular outputs
Every value in a key/value pair of a RecordOutput is now rendered as a
markup string.

Every row value of a TabularOutput is now also rendered as a markup
string. The following existing uses of TabularOutput are updated to use
markup strings instead of creating ANSI escape codes themselves:

* the unread mail indicator in ListMailCommand
* the names of knocked out players in GrappleEngine status
* the wound count of knocked out players in GrappleEngine status

These changes also implicitly mean that the sendMessage and
sendMessageToCaller API methods support markup strings for "record" and
"table" messages. Support for ordinary messages is coming shortly.